refactor(vulkan): rename vulkan/resources directory to vulkan/buffers for consistency
- Rename vulkan/resources directory to vulkan/buffers to match existing buffers directory structure
- Update all include paths from vulkan/resources to vulkan/buffers
- Maintain consistent naming convention across the codebase
- Ensure all files reference the new directory structure correctly

This change improves code organization by using consistent naming conventions.
The vulkan/buffers directory now aligns with the main buffers directory,
making the project structure more intuitive and easier to navigate.
