+++ include/vostok/graphics/backends/vulkan/buffers/vulkan_image.hpp
